fixup! Gives the runner access to RestrictionTracker

+++ sdks/java/core/src/test/java/org/apache/beam/sdk/transforms/SplittableDoFnTest.java
@@ -67,8 +67,8 @@ public class SplittableDoFnTest {
   static class PairStringWithIndexToLength extends DoFn<String, KV<String, Integer>> {
     @ProcessElement
     public ProcessContinuation process(ProcessContext c, OffsetRangeTracker tracker) {
-      for (int i = tracker.currentRestriction().getFrom(); tracker.tryClaim(i); ++i) {
-        c.output(KV.of(c.element(), i));
+      for (long i = tracker.currentRestriction().getFrom(); tracker.tryClaim(i); ++i) {
+        c.output(KV.of(c.element(), (int) i));
         if (i % 3 == 0) {
           return resume();
         }
@@ -202,8 +202,8 @@ public class SplittableDoFnTest {
     @ProcessElement
     public ProcessContinuation processElement(ProcessContext c, OffsetRangeTracker tracker) {
       int[] blockStarts = {-1, 0, 12, 123, 1234, 12345, 34567, MAX_INDEX};
-      int trueStart = snapToNextBlock(tracker.currentRestriction().getFrom(), blockStarts);
-      int trueEnd = snapToNextBlock(tracker.currentRestriction().getTo(), blockStarts);
+      int trueStart = snapToNextBlock((int) tracker.currentRestriction().getFrom(), blockStarts);
+      int trueEnd = snapToNextBlock((int) tracker.currentRestriction().getTo(), blockStarts);
       for (int i = trueStart; i < trueEnd; ++i) {
         if (!tracker.tryClaim(blockStarts[i])) {
           return resume();
